fixed storybook following refactor of stories into top level `story` dir
